Day 5 - decent perf today
I had one bug in part2. I only tried to fix the order once.  If the
result was not actually fixed because of multiple competing ordering
rules, then it needed to be tried again.

Originally I thought the fix() function couldn't be deterministic based
on the problem description.  I still think it might not be for some
orderings. But I suppose the ordering rules given force the result to be
deterministic.

Part1: 24 minutes
       About 10 minutes to parse it, and 12 minutes to solve it.

Part2: 43 minutes
       About 10 minutes to solve it, and 9 minutes to find the bug