Day15 Part 2
I had a stupid bug where I was treating empty cells like moveable
cells, but succeeding when all cells to be moved were empty.
This was fine in Part 1.  But in part 2, when  I was moving multiple
cells per level, my code didn't notice that some of the cells were
empty, so it was trying to push them anyway. And this made empty cells
push boxes.

Blarg!  So many moves in this thing, that I only found it by
comparing my maps output to someone's working solution.

Then the fix was easy.  But I should have developed my own
tests to try to confirm my code worked.  Dumb!